docker/manager/requirements.txt
djangorestframework
djangorestframework-simplejwt
drf-yasg
gunicorn
psycopg2-binary
requests
Pillow
